Refactor to a single `DandiS3MultipartManager`
